Add different base predictors
