+++ mmda/models/base_predictors/base_lp_predictor.py
@@ -1,0 +1,17 @@
+from abc import abstractmethod
+from typing import Union, List, Dict, Any
+
+from mmda.types.names import *
+from mmda.types.document import Document
+from mmda.types.annotation import BoxGroup
+from mmda.models.base_predictors.base_predictor import BasePredictor
+
+
+class BaseLayoutParserPredictor(BasePredictor):
+    REQUIRED_BACKENDS = ["layoutparser"]
+    REQUIRED_DOCUMENT_FIELDS = [Pages, Images]
+
+    def __init__(self, model: Any, config: Any):
+
+        self.model = model
+        self.config = config
